Refactor: Rename `task.rs` to `refresh.rs`
