+++ src/oauth2/jwk/mod.rs
@@ -45,6 +45,5 @@
 pub mod cache;
 pub mod jwk;
 
-mod task;
-
+mod refresh;
 mod util;
